+++ fe/fe-core/src/main/java/com/starrocks/sql/common/MetaUtils.java
@@ -19,6 +19,7 @@ import com.google.common.collect.Maps;
 import com.starrocks.analysis.Expr;
 import com.starrocks.analysis.TableName;
 import com.starrocks.catalog.Database;
+import com.starrocks.catalog.InternalCatalog;
 import com.starrocks.catalog.Table;
 import com.starrocks.common.AnalysisException;
 import com.starrocks.common.ErrorCode;
@@ -26,7 +27,6 @@ import com.starrocks.common.ErrorReport;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.qe.OriginStatement;
 import com.starrocks.qe.SqlModeHelper;
-import com.starrocks.server.CatalogMgr;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.sql.analyzer.SemanticException;
 import com.starrocks.sql.ast.CreateMaterializedViewStmt;
@@ -79,17 +79,12 @@ public class MetaUtils {
     }
 
     public static Database getDatabase(ConnectContext session, TableName tableName) {
-        Database db;
-        if (CatalogMgr.isInternalCatalog(tableName.getCatalog())) {
-            db = session.getGlobalStateMgr().getDb(tableName.getDb());
-            if (db == null) {
-                throw new SemanticException("Database %s is not found", tableName.getDb());
-            }
-        } else {
-            db = session.getGlobalStateMgr().getMetadataMgr().getDb(tableName.getCatalog(), tableName.getDb());
-            if (db == null) {
-                throw new SemanticException("Database %s is not found", tableName.getCatalogAndDb());
-            }
+        if (Strings.isNullOrEmpty(tableName.getCatalog())) {
+            tableName.setCatalog(session.getCurrentCatalog());
+        }
+        Database db = session.getGlobalStateMgr().getMetadataMgr().getDb(tableName.getCatalog(), tableName.getDb());
+        if (db == null) {
+            throw new SemanticException("Database %s is not found", tableName.getCatalogAndDb());
         }
         return db;
     }
@@ -103,45 +98,27 @@ public class MetaUtils {
     }
 
     public static Table getTable(TableName tableName) {
-        if (CatalogMgr.isInternalCatalog(tableName.getCatalog())) {
-            Database db = GlobalStateMgr.getCurrentState().getDb(tableName.getDb());
-            if (db == null) {
-                throw new SemanticException("Database %s is not found", tableName.getDb());
-            }
-            Table table = db.getTable(tableName.getTbl());
-            if (table == null) {
-                throw new SemanticException("Table %s is not found", tableName.getTbl());
-            }
-            return table;
-        } else {
-            Table table = GlobalStateMgr.getCurrentState().getMetadataMgr().getTable(tableName.getCatalog(),
-                    tableName.getDb(), tableName.getTbl());
-            if (table == null) {
-                throw new SemanticException("Table %s is not found", tableName);
-            }
-            return table;
+        if (Strings.isNullOrEmpty(tableName.getCatalog())) {
+            tableName.setCatalog(InternalCatalog.DEFAULT_INTERNAL_CATALOG_NAME);
         }
+        Table table = GlobalStateMgr.getCurrentState().getMetadataMgr().getTable(tableName.getCatalog(),
+                tableName.getDb(), tableName.getTbl());
+        if (table == null) {
+            throw new SemanticException("Table %s is not found", tableName);
+        }
+        return table;
     }
 
     public static Table getTable(ConnectContext session, TableName tableName) {
-        if (CatalogMgr.isInternalCatalog(tableName.getCatalog())) {
-            Database db = session.getGlobalStateMgr().getDb(tableName.getDb());
-            if (db == null) {
-                throw new SemanticException("Database %s is not found", tableName.getDb());
-            }
-            Table table = db.getTable(tableName.getTbl());
-            if (table == null) {
-                throw new SemanticException("Table %s is not found", tableName.getTbl());
-            }
-            return table;
-        } else {
-            Table table = session.getGlobalStateMgr().getMetadataMgr().getTable(tableName.getCatalog(),
-                    tableName.getDb(), tableName.getTbl());
-            if (table == null) {
-                throw new SemanticException("Table %s is not found", tableName.toString());
-            }
-            return table;
+        if (Strings.isNullOrEmpty(tableName.getCatalog())) {
+            tableName.setCatalog(session.getCurrentCatalog());
         }
+        Table table = session.getGlobalStateMgr().getMetadataMgr().getTable(tableName.getCatalog(),
+                tableName.getDb(), tableName.getTbl());
+        if (table == null) {
+            throw new SemanticException("Table %s is not found", tableName.toString());
+        }
+        return table;
     }
 
     public static Table getTable(String catalogName, String dbName, String tableName) {
